Move `Serializer` to `query` package
