+++ rdbms/src/main/java/io/spine/server/storage/jdbc/query/SelectMessageByIdQuery.java
@@ -24,7 +24,6 @@ import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Message;
 import com.querydsl.sql.AbstractSQLQuery;
 import io.spine.server.storage.jdbc.DatabaseException;
-import io.spine.server.storage.jdbc.Serializer;
 
 import javax.annotation.Nullable;
 import java.sql.ResultSet;
