+++ rdbms/src/main/java/io/spine/server/storage/jdbc/record/InsertEntityRecordsBulkQuery.java
@@ -26,7 +26,7 @@ import io.spine.server.entity.EntityRecord;
 import io.spine.server.entity.storage.ColumnRecords;
 import io.spine.server.entity.storage.EntityRecordWithColumns;
 import io.spine.server.storage.jdbc.query.IdColumn;
-import io.spine.server.storage.jdbc.Serializer;
+import io.spine.server.storage.jdbc.query.Serializer;
 import io.spine.server.storage.jdbc.query.ColumnAwareWriteQuery;
 import io.spine.server.storage.jdbc.query.Parameter;
 import io.spine.server.storage.jdbc.query.Parameters;
