+++ rdbms/src/main/java/io/spine/server/storage/jdbc/record/WriteEntityQuery.java
@@ -23,7 +23,7 @@ package io.spine.server.storage.jdbc.record;
 import com.google.common.base.Functions;
 import io.spine.server.entity.storage.ColumnRecords;
 import io.spine.server.entity.storage.EntityRecordWithColumns;
-import io.spine.server.storage.jdbc.Serializer;
+import io.spine.server.storage.jdbc.query.Serializer;
 import io.spine.server.storage.jdbc.query.Parameter;
 import io.spine.server.storage.jdbc.query.Parameters;
